endpoint_route_handler: add new tool model
