+++ endpoint_route_handler/models/__init__.py
@@ -1,3 +1,4 @@
 from . import endpoint_route_sync_mixin
 from . import endpoint_route_handler
+from . import endpoint_route_handler_tool
 from . import ir_http
