+++ endpoint_route_handler/models/endpoint_route_sync_mixin.py
@@ -81,7 +81,7 @@ class EndpointRouteSyncMixin(models.AbstractModel):
 
     def _register_hook(self):
         super()._register_hook()
-        if not self._abstract:
+        if not self._abstract and not self._transient:
             # Ensure existing active records are loaded at startup.
             # Pass `init` to bypass routing map refresh
             # since this piece of code runs only when the model is loaded.
